requirements.txt
faster-whisper
sounddevice
numpy
scipy
pynput
pystray
Pillow
httpx
customtkinter
